+++ src/handler/file.ts
@@ -7,14 +7,14 @@ import { mkdirpSync } from 'fs-extra';
 handlers.file = {
     text: {
         inputs: ['path', 'text'],
-        fn: (path: string, text: string) => {
+        fn: ([path, text]: [string, string]) => {
             mkdirpSync(path);
             writeFileSync(path, text, 'utf8');
         },
     },
     binary: {
         inputs: ['path', 'buffer'],
-        fn: (path: string, buffer: Buffer) => {
+        fn: ([path, buffer]: [string, Buffer]) => {
             mkdirpSync(path);
             writeFileSync(path, buffer, 'binary');
         },