+++ src/handler/to.ts
@@ -6,10 +6,10 @@ import Spider from '../core/Sipder';
 handlers.to = {
     urlQueue: {
         inputs: ['spider', 'urls'],
-        fn: (spider: Spider, urls: string[]) =>
+        fn: ([spider, urls]: [Spider, string[]]) =>
             _.each(urls, url => spider.urls.enqueue(url)),
     },
     console: {
-        fn: (data: string[]) => console.dir(data),
+        fn: ([data]: [string[]]) => console.dir(data),
     },
 }