+++ src/handler/urls.ts
@@ -5,6 +5,6 @@ import * as _ from 'lodash';
 handlers.urls = {
     resolve: {
         inputs: ['url', 'urls'], output: 'urls',
-        fn: (host: string, rUrls: string[]) => _.map(rUrls, link => host + link),
+        fn: ([host, rUrls]: [string, string[]]) => _.map(rUrls, link => host + link),
     },
 }