+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.1.4
+* Change handler arguments;
+* Retry for requesting failure.
+
 ## 0.1.2
 * Better Spider extend functions.